feat: refactor BlogFrontpage and Paginated components for improved structure and functionality
- Updated BlogFrontpage to utilize Section component for better semantic structure.
- Modified post fetching logic to allow for external posts prop, enhancing flexibility.
- Refactored Paginated component to import List from LatestArticles for consistency in article presentation.
